Support dynamic default values
Typer allows passing a `default_factory` to options and arguments, which
is a callable that will be used at runtime to determine the default
value. The Textual application created by Trogon crashes when this is
used, because it tries to convert the str representation of the callable
function for display when it should be converting the result of calling
that function instead. This change ensures dynamic default values do not
crash the application, and has the effect of setting the default value
appropriately in the TUI.